Quantifies rectilinearity based on image PSD.
First Commit. This repository stores code that quantifies an image’s
PSD via a 2D fft mapping. The mapping patterns are fed into a 3 layer
neural network that classifies the images as rectilinear or not. The
images are label sensitive. This method of classification forces the
network to ignore high level abstractions of the image (such as eyes)
and focuses on the frequency data which I hypothesize to be the key to
quantifying rectilinearity. The network generalizes very well to PPA
activations.